requirements.txt
customtkinter==5.2.2
pillow==11.0.0
python-dotenv==1.0.1
pydantic==2.10.1
watchdog==6.0.0
darkdetect==0.8.0
loguru==0.7.2
rich==13.9.4

# Dev dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.13.0 